# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ZipLink API Documentation",
        "description": "Высокопроизводительный сервис сокращения ссылок и сбора аналитики",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "Основной API сервер"
        }
    ],
    "paths": {
...
        "/{code}/guard": {
            "post": {
                "tags": [
                    "Ссылки"
                ],
                "summary": "Получение URL для ссылки, защищенной паролем",
                "operationId": "1fb55b9d4692101881b1f4cf040681c6",
                "parameters": [
                    {
                        "name": "code",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "1C"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "password"
                                ],
                                "properties": {
                                    "password": {
                                        "type": "string",
                                        "example": "secret123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Пароль верный, возврат URL",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "string",
                                            "example": "https://example.com/long-page"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Неверный пароль"
                    },
                    "404": {
                        "description": "Ссылка не найдена или истекла"
                    }
                }
            }
        },
...
        "/link/{id}/password": {
            "patch": {
                "tags": [
                    "Управление ссылками"
                ],
                "summary": "Установка или смена пароля ссылки",
                "operationId": "21b139ed82592e6f4c3aeaea3438c6e8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "password": {
                                        "type": "string",
                                        "example": "newpass",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Пароль обновлен"
                    },
                    "403": {
                        "description": "Нет прав доступа"
                    },
                    "404": {
                        "description": "Ссылка не найдена"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/link/{id}/expired": {
            "patch": {
                "tags": [
                    "Управление ссылками"
                ],
                "summary": "Установка или продление срока жизни ссылки",
                "operationId": "7fdfeac5407ddc97bf003d2d6a559105",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "expired_at": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2026-11-20T12:00:00Z",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Срок жизни обновлен"
                    },
                    "403": {
                        "description": "Нет прав"
                    },
                    "404": {
                        "description": "Ссылка не найдена"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/link/{id}/active": {
            "patch": {
                "tags": [
                    "Управление ссылками"
                ],
                "summary": "Включение или отключение активности ссылки",
                "operationId": "d010d8e44abc309a047d2bb5d1f5ab7c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "is_active"
                                ],
                                "properties": {
                                    "is_active": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Статус изменен"
                    },
                    "403": {
                        "description": "Нет прав"
                    },
                    "404": {
                        "description": "Ссылка не найдена"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/link/{id}": {
            "delete": {
                "tags": [
                    "Управление ссылками"
                ],
                "summary": "Удаление ссылки",
                "operationId": "24f54a579ec516f2a9fe3dabe5f0704b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ссылка удалена"
                    },
                    "403": {
                        "description": "Нет прав"
                    },
                    "404": {
                        "description": "Ссылка не найдена"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Введите токен в формате: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}